chore: clean up tests from previous pr
