+++ tests/devices/test_local_channel.py
@@ -2,6 +2,7 @@
 
 import asyncio
 import json
+import logging
 from collections.abc import Generator
 from unittest.mock import AsyncMock, Mock, patch
 
@@ -149,10 +150,11 @@ async def test_successful_command_response(local_channel: LocalChannel, mock_loo
 
 async def test_message_decode_error(local_channel: LocalChannel, caplog: pytest.LogCaptureFixture) -> None:
     """Test handling of message decode errors."""
-    local_channel._data_received(b"invalid_payload")
-    await asyncio.sleep(0.01)  # yield
+    with caplog.at_level(logging.WARNING):
+        local_channel._data_received(b"invalid_payload")
+        await asyncio.sleep(0.01)  # yield
 
-    warning_records = [record for record in caplog.records if record.levelname == "WARNING"]
+    warning_records = caplog.records
     assert len(warning_records) == 1
     assert "Failed to decode message" in warning_records[0].message
